+++ resources/constants.py
@@ -190,7 +190,7 @@ class Constants:
         self.set_content_caching = False  #    Set Content Caching
         self.allow_nvme_fixing = True  #       Allow NVMe Kernel Space Patches
         self.disable_xcpm = False  #           Disable XCPM (X86PlatformPlugin.kext)
-        self.root_patcher_succeded = False  #  Determine if root patcher succeeded
+        self.root_patcher_succeeded = False  #  Determine if root patcher succeeded
         self.booted_oc_disk = None  #          Determine current disk OCLP booted from
         self.start_build_install = False  #    Determine if build install should be started
         self.host_is_non_metal = False  #      Determine if host is non-metal (ie. enable UI hacks)